requirements.txt
aniso8601==4.0.1
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.6
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
python-dotenv==0.10.0
pytz==2018.7
six==1.12.0
Werkzeug==0.14.1
